Promote runtime materialization helpers
